@@ -1,6 +1,8 @@
 package resources
 
 import (
+	"time"
+
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
@@ -27,7 +29,6 @@ func ListAutoscalingGroups(s *session.Session) ([]Resource, error) {
 			}
 			return !lastPage
 		})
-
 	if err != nil {
 		return nil, err
 	}
@@ -65,7 +66,7 @@ func (asg *AutoScalingGroup) Properties() types.Properties {
 		properties.SetTag(tag.Key, tag.Value)
 	}
 
-	properties.Set("CreatedTime", asg.group.CreatedTime)
+	properties.Set("CreatedTime", asg.group.CreatedTime.Format(time.RFC3339))
 	properties.Set("Name", asg.group.AutoScalingGroupName)
 
 	return properties
