@@ -72,7 +72,7 @@ func (r *WAFRegionalRateBasedRulePredicate) Remove() error {
 		RuleId:      r.ruleID,
 		RateLimit:   r.rateLimit,
 		Updates: []*waf.RuleUpdate{
-			&waf.RuleUpdate{
+			{
 				Action:    aws.String("DELETE"),
 				Predicate: r.predicate,
 			},
